+++ .vscode/settings.json
@@ -28,6 +28,7 @@
         "rdsadapter",
         "rdsdb",
         "rdsserv",
+        "savingsgoals",
         "sdras",
         "seccomp",
         "Segoe",
