+++ internal/adapters/primary/web/app/accounts/handler.go
@@ -16,11 +16,6 @@ const (
 	CreditAccountsRoute = "/accounts/credits"
 	// NOTE: Savings accounts are basically normal accounts but their balance go to fulfill Savings Goals.
 	SavingsAccountsRoute = "/accounts/savings"
-	// NOTE: Saving Goals are a completely different data structure from accounts but it depends on accounts,
-	// specifically savings accounts. As the balance changes in the different savings account so does the completion in
-	// each savings goals. This goals contains an priority value order so the lower the priority value the earlier the
-	// the goal is fullfil.
-	SavingsGoalsAccountsRoute = "/accounts/savings_goals"
 	// NOTE: External are basically transaction categories but in this case are easier administer.
 	ExternalAccountsRoute = "/accounts/external"
 )
@@ -31,7 +26,6 @@ type Handler interface {
 	LoanAccountsHandlerFunc(c echo.Context) error
 	CreditAccountsHandlerFunc(c echo.Context) error
 	SavingsAccountsHandlerFunc(c echo.Context) error
-	SavingsGoalsHandlerFunc(c echo.Context) error
 	ExternalAccountsHandlerFunc(c echo.Context) error
 }
