--- internal/adapters/primary/web/app/views/accounts/htmx_savinggoals.templ
@@ -1,7 +1,0 @@
-package accounts
-
-templ SavingGoals[T SavingGoal](goals []T) {
-	for _, goal := range goals {
-		@savingGoal(goal)
-	}
-}
