+++ internal/adapters/primary/web/app/views/components/marquee_text.templ
@@ -1,0 +1,5 @@
+package components
+
+templ MarqueeHeading5(text string) {
+	<h5 class="text-lg inline-block w-fit whitespace-nowrap">{ text }</h5>
+}
