+++ internal/adapters/primary/web/app/views/savingsgoals/htmx_list.templ
@@ -1,0 +1,7 @@
+package savingsgoals
+
+templ HTMXList[Model SavingGoal](goals []Model) {
+	for i := 0; i < len(goals); i++ {
+		@savingsGoal(goals[i])
+	}
+}
